+++ engine/src/scene/grid_map.hpp
@@ -18,13 +18,13 @@ public:
     // 获取经过这个格子的消耗, < 0 表示不可行
     int8_t get_pass_cost(int32_t x, int32_t y) const;
     // 设置地图信息
-    bool set(int32_t id, uint16_t width, uint16_t height);
+    bool set(int32_t id, int32_t width, int32_t height);
     // 填充地图信息
-    bool fill(uint16_t x, uint16_t y, int8_t cost);
+    bool fill(int32_t x, int32_t y, int8_t cost);
 
     // 获取地图宽高
-    uint16_t get_width() const { return _width; }
-    uint16_t get_height() const { return _height; }
+    int32_t get_width() const { return _width; }
+    int32_t get_height() const { return _height; }
 
 private:
     int32_t _id;
@@ -33,7 +33,7 @@ private:
      * 一个格子可以表示64*64像素大小，或者32*32，具体看策划要求
      * 65535*32 = 2097120像素，一般情况下，uint16_t足够大了
      */
-    uint16_t _width;   // 地图的宽，格子坐标
-    uint16_t _height;  // 地图的长度，格子坐标
+    int32_t _width;   // 地图的宽，格子坐标
+    int32_t _height;  // 地图的长度，格子坐标
     int8_t *_grid_set; // 格子数据集合
 };
